NCIS {NCIS}	R Documentation

NCIS

Description

Estimates normalizing factor between a ChIP sample and a control/input sample

Usage

NCIS(chip.data, input.data, data.type = c("MCS", "BED", "AlignedRead"), frag.len = 200, min.binsize = 100, max.binsize = 20000, binsize.shift = 100, min.stop.binsize = 100, chr.vec = NULL, chr.len.vec = NULL, quant=0.75)

Arguments

chip.data
ChIP data.

input.data
control data.

data.type
"MCS", "AlignedRead" or "BED".

frag.len
average fragment length. Default 200 bp.

min.binsize
minimum of binsize to search.

max.binsize
maximum of binsize to search.

binsize.shift
the threshold of binsize after which the normalization factor is computed as the average of two estimates, one on regular bins and the other on bins shifed half binsize.

min.stop.binsize
minimum of binsize to use (stop).

chr.vec
vector of chromosomes in the data. Only reads in chr.vec are considered for normalization purpose.

chr.len.vec
vector of chromosome lengths corresponding to chr.vec

quant
quantile to start the searching for total threshold. Default 0.75.

Details

data.type

MCS Minimum ChIP-Seq format. data.frame with fields: chr (factor), pos (integer) and strand (factor, "+" and "-"). pos is 5' location. This is different from eland default which use 3' location for reverse strand.

AlignedRead from Bioconductor ShortRead package (with support of commonly used formats, including Eland, MAQ, Bowtie, SOAP and BAM).

BED with at least first 6 fields (chrom, start, end, name, score and strand), http://genome.ucsc.edu/FAQ/FAQformat.html#format1.

Value

A list with following components:

est
the estimated normalizing factor.

binsize.est
the estimated binsize.

r.seq.depth
sequencing depth ratio.

pi0
the estimated proportion of background reads among ChIP sample.

Author(s)

Kun Liang, kliang@stat.wisc.edu

References

Liang, K and Keles, S (2012). Normalization of ChIP-seq data with control, to appear.

Examples

data("SEG3")
res <- NCIS(chip.data, input.data, data.type="MCS")
res

[Package NCIS version 1.0.1 Index]